requirements.txt
discord.py
python-dotenv
flask
gunicorn
websockets
requests
